refactor: migrate getRevenueShare utility from licenseTermsHelper to royalty, update imports and tests accordingly
