+++ packages/core-sdk/src/resources/ipAsset.ts
@@ -105,7 +105,6 @@ import { handleError } from "../utils/errors";
 import { contractCallWithFees } from "../utils/feeUtils";
 import { generateOperationSignature } from "../utils/generateOperationSignature";
 import { getIpMetadataForWorkflow } from "../utils/getIpMetadataForWorkflow";
-import { getRevenueShare } from "../utils/licenseTermsHelper";
 import { PILFlavor } from "../utils/pilFlavor";
 import { handleMulticall } from "../utils/registrationUtils/registerHelper";
 import {
@@ -122,7 +121,7 @@ import {
   transferDistributeRoyaltyTokensRequest,
   transformRegistrationRequest,
 } from "../utils/registrationUtils/transformRegistrationRequest";
-import { royaltyPolicyInputToAddress } from "../utils/royalty";
+import { getRevenueShare, royaltyPolicyInputToAddress } from "../utils/royalty";
 import { validateAddress } from "../utils/utils";
 
 export class IPAssetClient {
