+++ packages/core-sdk/src/resources/license.ts
@@ -47,9 +47,8 @@ import { Erc20Spender } from "../types/utils/wip";
 import { calculateLicenseWipMintFee, predictMintingLicenseFee } from "../utils/calculateMintFee";
 import { handleError } from "../utils/errors";
 import { contractCallWithFees } from "../utils/feeUtils";
-import { getRevenueShare } from "../utils/licenseTermsHelper";
 import { PILFlavor } from "../utils/pilFlavor";
-import { royaltyPolicyInputToAddress } from "../utils/royalty";
+import { getRevenueShare, royaltyPolicyInputToAddress } from "../utils/royalty";
 import { waitForTxReceipt } from "../utils/txOptions";
 import { validateAddress } from "../utils/utils";
 import { validateLicenseConfig } from "../utils/validateLicenseConfig";
