+++ packages/core-sdk/src/utils/royalty.ts
@@ -2,6 +2,8 @@ import { Address } from "viem";
 
 import { chain, validateAddress } from "./utils";
 import { royaltyPolicyLapAddress, royaltyPolicyLrpAddress } from "../abi/generated";
+import { MAX_ROYALTY_TOKEN } from "../constants/common";
+import { RevShareType } from "../types/common";
 import { SupportedChainIds } from "../types/config";
 import { NativeRoyaltyPolicy, RoyaltyPolicyInput } from "../types/resources/royalty";
 
@@ -23,4 +25,19 @@ export const royaltyPolicyInputToAddress = (
       address = validateAddress(input);
   }
   return address;
+};
+
+export const getRevenueShare = (
+  revShare: number | string,
+  type: RevShareType = RevShareType.COMMERCIAL_REVENUE_SHARE,
+): number => {
+  const revShareNumber = Number(revShare);
+  if (isNaN(revShareNumber)) {
+    throw new Error(`${type} must be a valid number.`);
+  }
+
+  if (revShareNumber < 0 || revShareNumber > 100) {
+    throw new Error(`${type} must be between 0 and 100.`);
+  }
+  return (revShareNumber / 100) * MAX_ROYALTY_TOKEN;
 };
