+++ packages/core-sdk/test/unit/utils/royalty.test.ts
@@ -5,7 +5,7 @@ import {
   royaltyPolicyLapAddress,
   royaltyPolicyLrpAddress,
 } from "../../../src";
-import { royaltyPolicyInputToAddress } from "../../../src/utils/royalty";
+import { getRevenueShare, royaltyPolicyInputToAddress } from "../../../src/utils/royalty";
 import { aeneid, mockAddress } from "../mockData";
 
 describe("royaltyPolicyInputToAddress", () => {
@@ -33,5 +33,25 @@ describe("royaltyPolicyInputToAddress", () => {
   it("should return the correct address for a native LRP with a chainId", () => {
     const address = royaltyPolicyInputToAddress(NativeRoyaltyPolicy.LRP, "aeneid");
     expect(address).to.equal(royaltyPolicyLrpAddress[aeneid]);
+  });
+});
+
+describe("getRevenueShare", () => {
+  it("should throw error when call getRevenueShare given revShare is not a number", () => {
+    expect(() => getRevenueShare("not a number")).to.throw(
+      "CommercialRevShare must be a valid number.",
+    );
+  });
+
+  it("should throw error when call getRevenueShare given revShare is less than 0", () => {
+    expect(() => getRevenueShare(-1)).to.throw("CommercialRevShare must be between 0 and 100.");
+  });
+
+  it("should throw error when call getRevenueShare given revShare is greater than 100", () => {
+    expect(() => getRevenueShare(101)).to.throw("CommercialRevShare must be between 0 and 100.");
+  });
+
+  it("should return correct value when call getRevenueShare given revShare is 10", () => {
+    expect(getRevenueShare(10)).to.equal(10000000);
   });
 });
